requirements.txt
google-genai
yfinance
google-auth-oauthlib
google-api-python-client
requests
pandas